+++ packages/date-picker/src/DatePicker/DatePickerMenu/DatePickerMenuSelect/DatePickerMenuSelectYear.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import range from 'lodash/range';
 
+import { setUTCYear } from '@leafygreen-ui/date-utils';
 import { cx } from '@leafygreen-ui/emotion';
 import { Option, Select } from '@leafygreen-ui/select';
 
@@ -13,7 +14,7 @@ import {
 } from '../DatePickerMenu.styles';
 
 interface DatePickerMenuSelectYearProps {
-  onChange: (v: string) => void;
+  updateMonth: (newMonth: Date) => void;
 }
 
 /**
@@ -21,18 +22,23 @@ interface DatePickerMenuSelectYearProps {
  * @internal
  */
 export const DatePickerMenuSelectYear = ({
-  onChange,
+  updateMonth,
 }: DatePickerMenuSelectYearProps) => {
   const { setIsSelectOpen, min, max } = useSharedDatePickerContext();
   const { month } = useDatePickerContext();
   const yearOptions = range(min.getUTCFullYear(), max.getUTCFullYear() + 1);
+
+  const handleYearOnChange = (value: string) => {
+    const newMonth = setUTCYear(month, Number(value));
+    updateMonth(newMonth);
+  };
 
   return (
     <Select
       {...selectElementProps}
       aria-label="Select year"
       value={month.getUTCFullYear().toString()}
-      onChange={onChange}
+      onChange={handleYearOnChange}
       className={cx(selectTruncateStyles, selectInputWidthStyles)}
       onEntered={() => setIsSelectOpen(true)}
       onExited={() => setIsSelectOpen(false)}
