+++ packages/date-picker/src/shared/constants.ts
@@ -82,3 +82,5 @@ export const selectElementProps = {
   // there is currently no way to pass a ref into the Select portal to use in backdrop "foreground"
   usePortal: false,
 } as const;
+
+export const yearFirstFormats = ['iso8601'];
